fix structure fs
