+++ controllers/contacts/listContacts.js
@@ -1,0 +1,9 @@
+const sendSuccessRes = require('../../helpers/sendSuccessRes')
+const contactsOperations = require('../../model/index')
+
+const listContacts = async(req, res) => {
+  const contacts = await contactsOperations.listContacts()
+  sendSuccessRes(res, { contacts })
+}
+
+module.exports = listContacts
