+++ model/contacts/updateContacts.js
@@ -2,9 +2,9 @@
 const fs = require('fs/promises')
 const path = require('path')
 
-const contactsPath = path.join(__dirname, 'contacts.json')
+const contactsPath = path.join(__dirname, '..', 'db', 'contacts.json')
 
 const updateContacts = async contacts => {
-  await fs.writeFile(contactsPath, JSON.stringify(contacts))
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, ' '))
 }
 module.exports = updateContacts
